third_party/blink/web_tests/platform/generic/fast/dom/offset-position-writing-modes-expected.txt

PASS document.getElementById("vertical").offsetLeft is 65
PASS successfullyParsed is true

TEST COMPLETE
